+++ processor/database/stories_db.py
@@ -110,18 +110,17 @@ def recent_stories(project_id: int, above_threshold: bool, limit: int = 5) -> Li
     """
     earliest_date = dt.date.today() - dt.timedelta(days=7)
     session = Session()
-    q = session.execute(
-    select(Story)
-    .where(
-        (Story.project_id == project_id) &
-        (Story.above_threshold == above_threshold) &
-        (Story.published_date > earliest_date)
-    )
-    .order_by(func.random())
-    .limit(limit)
-    .all())#[updated]
-    stories = [s for s in q]
-    return stories
+    result = session.execute(
+        select(Story)
+        .where(
+            (Story.project_id == project_id) &
+            (Story.above_threshold == above_threshold) &
+            (Story.published_date > earliest_date)
+        )
+        .order_by(func.random())
+        .limit(limit)
+        )
+    return result.scalars().all()
 
 
 def _stories_by_date_col(column_name: str, project_id: int = None, platform: str = None, above_threshold: bool = None,
@@ -162,7 +161,7 @@ def _run_query(query: str) -> List:
     with processor.engine.begin() as connection:
         result = connection.execute(text(query))
         for row in result:
-            data.append(dict(row))
+            data.append(row._mapping)
     return data