+++ CHANGELOG.md
@@ -3,6 +3,18 @@ Change Log
 
 Here is a history of what was changed in each version. 
 
+### v3.4.4
+
+* Better error handling for Newscatcher, longer window of days to search within
+
+### v3.4.3
+
+* More work on SQLalchemy v2 integration to fix web dashboard
+
+### v3.4.2
+
+* Fix query results architecture to work on dashboard errors
+
 ### v3.4.1
 
 * Upgrade sklearn to try and fix deploy compilation crash